Fix bug: onUpdate() was not always called with the complete object when the location changed
If the object moved into the bounds of an existing listener, the
listener was only sent the changed properties (or more accurately, the
properties that .updateObject() got) and not the whole
object as it should be.

Renamed the variables within .updateObject() to make it clearer which
object is which.

This could have prevented if the model was simpler: for example, if
objects could not move at all. In that case, movement would be
represented by deleting an existing object and then creating a new
one in the new location.

That might not be the best solution for clients that want to be able to
animate moving objects in a simple way, though.